requirements.txt

ipykernel==6.22.0
matplotlib-inline==0.1.6
nbconvert==7.3.1
notebook==6.5.4
numpy==1.24.3
pyro-ppl==1.5.2
torch==1.7.1
torchvision==0.8.2
scikit-learn
matplotlib